+++ src/main/kotlin/org/mockito4kotlin/annotation/KMockitoAnnotations.kt
@@ -1,7 +1,8 @@
 /*
+ *
  * The MIT License
  *
- *   Copyright (c) 2017-2018 Wilhelm Schulenburg
+ *   Copyright (c) 2017-2021 Wilhelm Schulenburg
  *   Copyright (c) 2007 Mockito contributors
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -21,6 +22,7 @@
  *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *  THE SOFTWARE.
+ *
  */
 
 package org.mockito4kotlin.annotation
@@ -38,7 +40,7 @@ import kotlin.reflect.jvm.isAccessible
  * * Makes code easier to read.
  * * Separate mock initialization from test code.
  * * Works like @[org.mockito.Mock] annotation of Mockito.
- * * With [KMock] you are able to use Mockitos Annotation by data classes and classes with delegates.
+ * * With [KMock] you are able to use Mockito's Annotation by data classes and classes with delegates.
  *
  * See examples in [org.mockito.Mock]
  *
@@ -54,6 +56,7 @@ import kotlin.reflect.jvm.isAccessible
  * @property name Name of this mock.
  * @property answer Default answers to interactions.
  * @property serializable Mock is serializable.
+ * @property lenient Mock will be lenient [org.mockito.MockSettings.lenient].
  *
  */
 @Target(AnnotationTarget.PROPERTY, AnnotationTarget.FIELD)
@@ -62,10 +65,11 @@ annotation class KMock(val extraInterfaces: Array<KClass<out Any>> = [],
                        val stubOnly: Boolean = false,
                        val name: String = "",
                        val answer: Answers = Answers.RETURNS_DEFAULTS,
-                       val serializable: Boolean = false)
+                       val serializable: Boolean = false,
+                       val lenient: Boolean = false)
 
 /**
- * Allows shorthand [com.nhaarman.mockito_kotlin.KArgumentCaptor] creation on fields,
+ * Allows shorthand [com.nhaarman.mockitokotlin2.KArgumentCaptor] creation on fields,
  * properties or property's backing fields.
  *
  * * Makes code easier to read.
@@ -105,71 +109,53 @@ annotation class KCaptor
  * @see org.mockito.Captor
  * @see org.mockito.InjectMocks
  */
-class KMockitoAnnotations {
-    companion object {
-        private val mockedPropertiesAssigned = MockPropertyCollector()
+object KMockitoAnnotations {
+    private val mockedAssignedProperties = MockPropertyCollector()
 
-        /**
-         * Initializes objects annotated with [Mock], [KMock], [Spy], [Captor], [KCaptor] and [InjectMocks].
-         *
-         * @param anyWithMocks instance with Mockannotations to be initialized.
-         *
-         */
-        fun initMocks(anyWithMocks: Any) {
-            mockedPropertiesAssigned.reset()
-            val properties = anyWithMocks::class.memberProperties
+    /**
+     * Initializes objects annotated with [Mock], [KMock], [Spy], [Captor], [KCaptor] and [InjectMocks].
+     *
+     * @param anyInstanceWithMocks instance with MockAnnotations to be initialized.
+     */
+    fun initMocks(anyInstanceWithMocks: Any) {
+        mockedAssignedProperties.reset()
+        val properties = anyInstanceWithMocks::class.memberProperties
 
-            processAnnotations(anyWithMocks, properties, ::isMockito4KotlinAnnotation)
-            registerInlinedMockDeclaration(anyWithMocks, properties)
-            processAnnotations(anyWithMocks, properties) { it is InjectMocks }
-        }
+        processAnnotations(anyInstanceWithMocks, properties, ::isMockito4KotlinAnnotation)
+        registerInlinedMockDeclaration(anyInstanceWithMocks, properties)
+        processAnnotations(anyInstanceWithMocks, properties) { it is InjectMocks }
+    }
 
-        private fun processAnnotations(anyWithMocks: Any, properties: Collection<KProperty<*>>, predicate: Predicate<Annotation>) {
-            properties.filter { it.allAnnotations().any(predicate) }
-                .forEach { property ->
-                    AnnotationEngineFactory.create(asAnnotation(property, predicate)).apply {
-                        inject(this@Companion.mockedPropertiesAssigned)
-                        process(anyWithMocks, property)
-                    }
+    private fun processAnnotations(anyInstanceWithMocks: Any, properties: Collection<KProperty<*>>, predicate: Predicate<Annotation>) {
+        properties.filter { it.allAnnotations().any(predicate) }
+            .forEach { property ->
+                AnnotationEngineFactory.create(asAnnotation(property, predicate)).apply {
+                    inject(mockedAssignedProperties)
+                    process(anyInstanceWithMocks, property)
                 }
+            }
+    }
+
+    private fun registerInlinedMockDeclaration(anyInstanceWithMocks: Any, properties: Collection<KProperty<*>>) {
+
+        fun isValueMocked(value: Any): Boolean {
+            val mockingDetails = Mockito.mockingDetails(value)
+            return mockingDetails.isMock || mockingDetails.isSpy
         }
 
-        private fun registerInlinedMockDeclaration(anyWithMocks: Any, properties: Collection<KProperty<*>>) {
-
-            fun isValueMocked(value: Any): Boolean {
-                val mockingDetails = Mockito.mockingDetails(value)
-                return mockingDetails.isMock || mockingDetails.isSpy
-            }
-
-            properties.forEach { property ->
-                property.isAccessible = true
-                property.getter.call(anyWithMocks)?.let { value ->
-                    if (isValueMocked(value)) {
-                        mockedPropertiesAssigned.register(property, value)
-                    }
+        properties.forEach { property ->
+            property.isAccessible = true
+            property.getter.call(anyInstanceWithMocks)?.let { value ->
+                if (isValueMocked(value)) {
+                    mockedAssignedProperties.register(property, value)
                 }
             }
         }
-
-        private fun asAnnotation(property: KProperty<*>, predicate: Predicate<Annotation>) =
-            property.allAnnotations().find(predicate)
-
     }
-}
 
-@Deprecated("This class is deprecated. Use 'KMockitoAnnotations' instead.",
-    replaceWith = ReplaceWith("KMockitoAnnotations")
-)
-class MockAnnotations {
-    companion object {
-        @Deprecated("This function is deprecated. Use 'KMockitoAnnotations.initMocks(Any)' instead.",
-            replaceWith = ReplaceWith("KMockitoAnnotations.initMocks(anyWithMocks)"),
-            level = DeprecationLevel.ERROR
-        )
-        fun initMocks(anyWithMocks: Any) {
-            KMockitoAnnotations.initMocks(anyWithMocks)
-        }
-    }
+    private fun asAnnotation(property: KProperty<*>, predicate: Predicate<Annotation>) =
+        property.allAnnotations().find(predicate)
+
 }
 
 /**