+++ src/main/kotlin/org/mockito4kotlin/annotation/engine/CaptorAnnotationEngine.kt
@@ -1,7 +1,8 @@
 /*
+ *
  * The MIT License
  *
- *   Copyright (c) 2017-2019 Wilhelm Schulenburg
+ *   Copyright (c) 2017-2021 Wilhelm Schulenburg
  *   Copyright (c) 2007 Mockito contributors
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -21,6 +22,7 @@
  *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *  THE SOFTWARE.
+ *
  */
 
 package org.mockito4kotlin.annotation.engine
@@ -37,13 +39,13 @@ import kotlin.reflect.jvm.isAccessible
 
 internal class CaptorAnnotationEngine : AbstractAnnotationEngine() {
 
-    override fun process(anyWithMocks: Any, property: KProperty<*>) {
+    override fun process(anyInstanceWithMocks: Any, property: KProperty<*>) {
         property.isAccessible = true
         checkImmutableProperties(property)
         checkNumberOfMockAnnotations(property)
         checkIsArgumentCaptor(property)
 
-        (property as KMutableProperty<*>).setter.call(anyWithMocks, createArgumentCaptor(property))
+        (property as KMutableProperty<*>).setter.call(anyInstanceWithMocks, createArgumentCaptor(property))
     }
 
     private fun createArgumentCaptor(property: KProperty<*>): Any =